# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing shared expenses, groups, and user authentication",
        "title": "Qashare API",
        "contact": {
            "name": "Pranaov S",
            "email": "qashare.contact@pranaovs.me"
        },
        "license": {
            "name": "AGPL-3.0",
            "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/expenses/": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new expense with splits for a group. The logged in user will be set as the AddedBy user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expenses"
                ],
                "summary": "Create a new expense",
                "parameters": [
                    {
                        "description": "Expense details with splits",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ExpenseDetails"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Expense successfully created with splits",
                        "schema": {
                            "$ref": "#/definitions/models.ExpenseDetails"
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST: Invalid request body, missing required fields, or no splits provided | INVALID_SPLIT: Split totals do not match expense amount or split validation failed",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "401": {
                        "description": "INVALID_TOKEN: Authentication token is missing, invalid, or expired",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "403": {
                        "description": "USERS_NOT_RELATED: The authenticated user is not a member of the specified group | USER_NOT_IN_GROUP: One or more users in the splits are not members of the group",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "404": {
                        "description": "GROUP_NOT_FOUND: The specified group does not exist",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal server error - unexpected database error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    }
                }
            }
        },
        "/v1/expenses/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed information about an expense including splits",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expenses"
                ],
                "summary": "Get expense details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Expense ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns expense details including all splits",
                        "schema": {
                            "$ref": "#/definitions/models.ExpenseDetails"
                        }
                    },
                    "401": {
                        "description": "INVALID_TOKEN: Authentication token is missing, invalid, or expired",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "403": {
                        "description": "USERS_NOT_RELATED: The authenticated user is not a member of the group this expense belongs to",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "404": {
                        "description": "EXPENSE_NOT_FOUND: The specified expense does not exist",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal server error - unexpected database error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing expense (requires being group admin or expense creator). Immutable fields will be ignored if included in the request body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expenses"
                ],
                "summary": "Update an expense",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Expense ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated expense details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ExpenseDetails"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns updated expense with all fields",
                        "schema": {
                            "$ref": "#/definitions/models.ExpenseDetails"
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST: Invalid request body or missing required fields | INVALID_SPLIT: No splits provided or split totals do not match expense amount",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "401": {
                        "description": "INVALID_TOKEN: Authentication token is missing, invalid, or expired",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "403": {
                        "description": "NO_PERMISSIONS: User is not the group admin or expense creator | USERS_NOT_RELATED: The authenticated user is not a member of the group | USER_NOT_IN_GROUP: One or more users in the splits are not members of the group",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "404": {
                        "description": "EXPENSE_NOT_FOUND: The specified expense does not exist",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal server error - unexpected database error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an expense (requires being group admin or expense creator)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expenses"
                ],
                "summary": "Delete an expense",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Expense ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "INVALID_TOKEN: Authentication token is missing, invalid, or expired",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "403": {
                        "description": "NO_PERMISSIONS: User is not the group admin or expense creator | USERS_NOT_RELATED: The authenticated user is not a member of the group",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "404": {
                        "description": "EXPENSE_NOT_FOUND: The specified expense does not exist",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal server error - unexpected database error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update specific fields of an expense. Only provided fields are updated, others remain unchanged. Immutable fields are automatically protected.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expenses"
                ],
                "summary": "Partially update an expense",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Expense ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Partial expense details (all fields optional except where validation requires)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ExpenseDetails"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns updated expense with all fields",
                        "schema": {
                            "$ref": "#/definitions/models.ExpenseDetails"
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST: Invalid request body or validation failed | INVALID_SPLIT: Split totals do not match expense amount",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "401": {
                        "description": "INVALID_TOKEN: Authentication token is missing, invalid, or expired",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "403": {
                        "description": "NO_PERMISSIONS: User is not the group admin or expense creator | USERS_NOT_RELATED: The authenticated user is not a member of the group | USER_NOT_IN_GROUP: One or more users in the splits are not members of the group",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "404": {
                        "description": "EXPENSE_NOT_FOUND: The specified expense does not exist",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal server error - unexpected database error",
                        "schema": {
                            "$ref": "#/definitions/apierrors.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apierrors.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "e.g., \"BAD_NAME\", \"INVALID_EMAIL\"",
                    "type": "string"
                },
                "message": {
                    "description": "Human-readable message",
                    "type": "string"
                }
            }
        },
...
        "models.ExpenseDetails": {
            "type": "object",
            "properties": {
                "added_by": {
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "integer"
                },
                "description": {
                    "description": "pointer because nullable in db",
                    "type": "string"
                },
                "expense_id": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "is_incomplete_amount": {
                    "type": "boolean"
                },
                "is_incomplete_split": {
                    "type": "boolean"
                },
                "latitude": {
                    "description": "pointer because nullable in db",
                    "type": "number"
                },
                "longitude": {
                    "description": "pointer because nullable in db",
                    "type": "number"
                },
                "splits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ExpenseSplit"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.ExpenseSplit": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "is_paid": {
                    "description": "\"paid\" or \"owes\"",
                    "type": "boolean"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}